+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.14.56] - 2021-01-11
 - Fix accumulated tokens registeration
+- Fix token accumulation
 
 ## [0.14.55] - 2021-01-10
 - Fix accumulated tokens registeration